SIGNALR messaging

+++ SnakeVS/Shared/SnakeFood.cs
@@ -6,8 +6,8 @@ namespace SnakeVS.Shared
     public class SnakeFood
     {
         [Key(0)]
-        public sbyte PositionX { get; set; }
+        public int PositionX { get; set; }
         [Key(1)]
-        public sbyte PositionY { get; set; }
+        public int PositionY { get; set; }
     }
 }
